backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
websockets==12.0

# Database dependencies
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Redis dependencies
redis==4.6.0
hiredis==2.2.3

# Celery dependencies (for v3.0 intelligence learning tasks)
celery[redis]==5.3.4

# Vector Database dependencies
qdrant-client==1.7.0

# AI/LLM dependencies
openai>=1.6.1
anthropic==0.7.7

# Trading dependencies
hyperliquid-python-sdk==0.20.0
python-binance==1.0.19

# HTTP client dependencies
httpx==0.25.2
aiohttp==3.9.1

# Authentication dependencies
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-dotenv==1.0.0

# Data processing dependencies
pandas==2.1.4
numpy==1.26.2

# Configuration dependencies
pydantic==2.5.2
pydantic-settings==2.1.0
email-validator==2.1.0
tenacity==8.2.3
pytz==2023.3

# Logging dependencies
loguru==0.7.2

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
PyJWT==2.8.0
